hehe

+++ game/moveable.cpp
@@ -1,8 +1,9 @@
 #include <SFML/Window.hpp>
 #include <SFML/Graphics.hpp>
 #include "moveable.h"
+#include <vector>
 
-Moveable::Moveable(const std::vector<int> &position, const std::vector<int> &size)
+Moveable::Moveable(const std::vector<float> &position, const std::vector<float> &size)
 {
     if(position.size()==3)
     {
